requirements.txt
opencv-python
numpy
torch
torchvision
sqlalchemy
pyyaml
streamlit
fastapi
uvicorn
python-multipart
